Document Editor Translation/backend/requirements.txt
fastapi==0.115.12
protobuf==4.25.3
pydantic==2.10.6
python-dotenv==1.1.0
uvicorn==0.34.0
google-genai==1.7.0